+++ packages/camera/camera_windows/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.0
+* Adds support for custom video output path in video recording.
+
 ## 0.2.6+5
 
 * Updates pigeon dev_dependency to ^27.3.2 for analyzer 14 compatibility.